refactor: move security-related class to a common module for reusability
