+++ gildongmu-common/src/main/java/codeit/common/exception/ErrorCode.java
@@ -7,8 +7,8 @@ import org.springframework.http.HttpStatus;
 @Getter
 @RequiredArgsConstructor
 public enum ErrorCode {
-    // domain
-    DOMAIN_NOT_FOUND(HttpStatus.BAD_REQUEST, "해당하는 도메인이 없습니다.");
+    // user
+    USER_NOT_FOUND(HttpStatus.BAD_REQUEST, "해당하는 유저가 없습니다.");
     private final HttpStatus httpStatus;
     private final String message;
 }
