+++ gildongmu-common/src/main/java/codeit/common/security/UserDetailsServiceImpl.java
@@ -1,6 +1,6 @@
-package codeit.api.security;
+package codeit.common.security;
 
-import codeit.api.exception.ErrorCode;
+import codeit.common.exception.ErrorCode;
 import codeit.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -11,11 +11,13 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class UserDetailsServiceImpl implements UserDetailsService {
+
     private final UserRepository userRepository;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return new UserDetailsImpl(userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException(ErrorCode.USER_NOT_FOUND.getMessage())));
+        return new UserPrincipal(userRepository.findByEmail(username)
+            .orElseThrow(
+                () -> new UsernameNotFoundException(ErrorCode.USER_NOT_FOUND.getMessage())));
     }
 }
